requirements.txt
boto3==1.34.131
python-dotenv
streamlit
